Added whitelist storage drivers per node type
The NodeType enum now has a list of allowed StorageDriverKinds. When
creating a new storage pool the controller performs a check whether the
given storage driver is allowed for that node-type or not.

Additionally added the new SWORDFISH_TARGET node-type.

+++ server/src/main/java/com/linbit/extproc/DaemonHandler.java
@@ -47,13 +47,16 @@ public class DaemonHandler
     {
         outProxy.expectShutdown();
         errProxy.expectShutdown();
-        if (force)
+        if (process != null)
         {
-            process.destroyForcibly();
-        }
-        else
-        {
-            process.destroy();
+            if (force)
+            {
+                process.destroyForcibly();
+            }
+            else
+            {
+                process.destroy();
+            }
         }
         outThread.interrupt();
         errThread.interrupt();
